# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Sistem Informasi Kepegawaian dan Status Kehadiran API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /roles/{id}/permissions:
    get:
      operationId: getRolePermissions
      tags: [roles]
      summary: Get permissions for a role
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Permissions assigned to role
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Permission"
  /roles/{id}/permissions/{permissionId}:
    post:
      operationId: assignPermissionToRole
      tags: [roles]
      summary: Assign a permission to a role
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: permissionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Permission assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
    delete:
      operationId: removePermissionFromRole
      tags: [roles]
      summary: Remove a permission from a role
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: permissionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Permission removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /inventory/items:
    get:
      operationId: getInventoryItems
      tags: [inventory]
      summary: Get all inventory items
      responses:
        "200":
          description: List of inventory items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
    post:
      operationId: createInventoryItem
      tags: [inventory]
      summary: Create new inventory item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInventoryItem"
      responses:
        "201":
          description: Item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
  /inventory/items/{id}:
    put:
      operationId: updateInventoryItem
      tags: [inventory]
      summary: Update inventory item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInventoryItem"
      responses:
        "200":
          description: Item updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
    delete:
      operationId: deleteInventoryItem
      tags: [inventory]
      summary: Delete inventory item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Item deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /inventory/transactions:
    get:
      operationId: getInventoryTransactions
      tags: [inventory]
      summary: Get inventory transactions
      parameters:
        - name: tipe
          in: query
          schema:
            type: string
            enum: [masuk, keluar]
        - name: itemId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryTransaction"
    post:
      operationId: createInventoryTransaction
      tags: [inventory]
      summary: Create inventory transaction (barang masuk/keluar)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInventoryTransaction"
      responses:
        "201":
          description: Transaction created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryTransaction"
  /inventory/transactions/{id}:
    delete:
      operationId: deleteInventoryTransaction
      tags: [inventory]
      summary: Delete inventory transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transaction deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Permission:
      type: object
      properties:
        id:
          type: integer
        resource:
          type: string
        action:
          type: string
        description:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - resource
        - action
        - createdAt
...
    InventoryItem:
      type: object
      properties:
        id:
          type: integer
        kodeBarang:
          type: string
        namaBarang:
          type: string
        kategori:
          type: string
          nullable: true
        satuan:
          type: string
          nullable: true
        stok:
          type: integer
        keterangan:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - kodeBarang
        - namaBarang
        - stok
        - createdAt
    CreateInventoryItem:
      type: object
      properties:
        kodeBarang:
          type: string
        namaBarang:
          type: string
        kategori:
          type: string
          nullable: true
        satuan:
          type: string
          nullable: true
        stok:
          type: integer
        keterangan:
          type: string
          nullable: true
      required:
        - kodeBarang
        - namaBarang
        - stok
    InventoryTransaction:
      type: object
      properties:
        id:
          type: integer
        itemId:
          type: integer
        item:
          $ref: "#/components/schemas/InventoryItem"
        tipe:
          type: string
          enum: [masuk, keluar]
        jumlah:
          type: integer
        tanggal:
          type: string
        keterangan:
          type: string
          nullable: true
        penanggungJawab:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - itemId
        - tipe
        - jumlah
        - tanggal
        - createdAt
    CreateInventoryTransaction:
      type: object
      properties:
        itemId:
          type: integer
        tipe:
          type: string
          enum: [masuk, keluar]
        jumlah:
          type: integer
        tanggal:
          type: string
        keterangan:
          type: string
          nullable: true
        penanggungJawab:
          type: string
          nullable: true
      required:
        - itemId
        - tipe
        - jumlah
        - tanggal
...